feat: Integrate Anthropic prompt caching into generation pipeline
Implements prompt caching support to achieve 60-80% reduction in input token costs
for Anthropic Claude models. The implementation separates static, cacheable content
(coding standards, FCS context) from dynamic, task-specific content.

Changes:
- Added CacheableClient interface extending Client with GenerateWithCache() method
- Created buildCodeGenerationPromptWithCache() in coder.go using PromptBuilder
- Created buildPlanningPromptWithCache() in planner.go for planning stage
- Modified GenerateFile() to auto-detect caching support and use when available
- Added cache metrics tracking to GenerationMetrics (hits, misses, tokens saved)
- Integrated cache performance logging with hit rate and cost savings
- Fixed linting issues (ineffective assignment, unnecessary conversions)

Technical Details:
- Static content (coding standards, FCS schema) marked as cacheable with 5m TTL
- Dynamic content (file paths, task details) sent as non-cached user messages
- Graceful fallback to standard generation for non-Anthropic providers
- Cache metrics automatically collected and logged after generation runs

Testing:
- All unit tests pass (prompt builder, cache metrics, config validation)
- All integration tests pass (generation flow with cache detection)
- Linting passes with 0 issues
- Code compiles successfully

Expected Impact:
- 60-80% reduction in input tokens for multi-file generation
- Automatic transparent integration with Anthropic clients
- Full backward compatibility with other LLM providers

Related: specs/003-performance-optimization/spec.md User Story 1

+++ pkg/llm/anthropic.go
@@ -241,18 +241,18 @@ func (c *anthropicClient) GenerateWithCache(ctx context.Context, messages []Cach
 
 		// Update cache metrics from usage
 		if response.Usage.CacheCreationInputTokens > 0 {
-			c.cacheMetrics.CacheCreationTokens += int64(response.Usage.CacheCreationInputTokens)
+			c.cacheMetrics.CacheCreationTokens += response.Usage.CacheCreationInputTokens
 			c.cacheMetrics.CacheMisses++
 		}
 		if response.Usage.CacheReadInputTokens > 0 {
-			c.cacheMetrics.CacheReadTokens += int64(response.Usage.CacheReadInputTokens)
+			c.cacheMetrics.CacheReadTokens += response.Usage.CacheReadInputTokens
 			c.cacheMetrics.CacheHits++
 		}
 		if response.Usage.InputTokens > 0 {
-			c.cacheMetrics.InputTokens += int64(response.Usage.InputTokens)
+			c.cacheMetrics.InputTokens += response.Usage.InputTokens
 		}
 		if response.Usage.OutputTokens > 0 {
-			c.cacheMetrics.OutputTokens += int64(response.Usage.OutputTokens)
+			c.cacheMetrics.OutputTokens += response.Usage.OutputTokens
 		}
 
 		return nil
